+++ releases.md
@@ -21,9 +21,20 @@ Changes are grouped as follows:
 
 ## [1.4.0-SNAPSHOT]
 
+### Added
+
+- Support for `ExtractionPipeline` and `ExtractionPipelineRun` so you can send extractor/pipeline observations and 
+ heartbeats to CDF.
+- Improved performance of `Relationships.list()` with the added support for partitions.
+- Support for including the `source` and `target` object of a `relationship` when using `list()` or `retrieve()`.
+- Improved performance of `Sequences.list()` with the added support for partitions.
+
 ### Fixed
 
 - More stability improvements to file binary downloads, in particular in situations with limited bandwidth. 
+- When trying to upload a `File` without a binary, the SDK could throw an exception if the binary was set to 
+ be a `URI`.
+- `Relationship.sourceType` was wrongly set equal to `Relationship.targetType`.
 
 ## [1.3.0] 2021-10-12